Create tracking information from campaign config
Use the campaign configuration to generate tracking information inside
the banners. During development, the tracking information is injected
dynamiccaly for the current banner. For the compiled banners, the
tracking information is baked into the Wikitext wrapper around the
JavaScript.

Putting the tracking information in the redundant data attribute
`data-trackingparams` solves the problem of missing tracking information
when the backend requests the banner code and looks for `piwik_*` URL
parameters.