Improve number formatting
Use `format-number` library instead of `numeral`: numeral is not only
about formatting numbers but is also parsing them and for doing math.

Add `format-number` to dependencies: Since this is a library that is
literally "baked into" the banner, I put it as a dependency instead of a
dev dependency.

Add `Mockdate` library to dev dependencies.