Fix missing compiled variables in mini banner
Due to the variables not being in the same file, postcss fails to load
them and does not replace the var() blocks with backwards compatible
raw values. By forcing postcss to not preserve the original variables, we
reduce the size of the output CSS and make sure that we notice this more
easily, as missing variables would now break on all browsers.

https://phabricator.wikimedia.org/T211099