Add local impression counting
The FundraisingFrontend needs data how many times a banner was displayed
and how many times the current banner was displayed. Previousy, this was
implemented in BannerFunctions, now it's a standalone class.

NOTE: The new implementation uses localStorage instead of cookies
because its API is more easy to work with. When using this in older
banners, please add a compatibility wrapper that converts the cookie
values into localStorage ones, to keep continuity.

The storage keys also change from
`centralnotice_banner_impression_count` and
`centralnotice_single_banner_impression_count` to
`fundraising.overallCount` and `fundraising.bannerCount`.