Add preact-based progress bar to desktop banner (#254)
Add getRemainingDays method to CampaignProjection
Add Progress Bar component for Desktop (might be turned into a mobile
one later).
Add TranslationContext for passing around translations.
Add TODOs for optimizing the Infobox component and reducing the amount
of code in banner_ctrl.js

The progress bar animation is trigged by a callback from the banner.
The progress bar animation is done via CSS transitions.

Add prop-types library and use it to describe the props of the progress
bar. This example should be extended to other class components that use properties.

The current Preact version (10.x) does not validate the propTypes
automatically, so we need to that in the constructor.