Set up wp.de Test 01
This test tests different breakpoints, which are stored in PostCSS
variables. Since there are no "dynamic imports" in postcss for the
variable files, the CSS files are duplicated.

THIS IS A HACK, NOT BEST PRACTICE!

We'll need to think about how to structure our styles independently from
the components, in a way that makes it obvious which components are
styled differently and in what way.