Implement and extract banner animation
Add the class BannerTransition class so it can be used
to slide in any banner to the page.
The animation is done via performant CSS transition in order
to get away from a solution with jQuery.
The BannerTransition has a onFinish property to be able to let the banner know e.g. when the progress bar animatio should start.
It is important for the banner-transition Element to be position: absolute to be visible properly.

The skins vector and minerva now have the same css animation style because the jQuery animation easing differs.