Finish banner style and layout
This commit makes the new preact banner look like the old desktop CTRL banner.
It adjusts the layout, missing content and styling.

Other decisions during the process:

The PCSS (styling) of each UI component in /shared/
now is done in the banner directories, because they
differ and alter too much and too often during the campaign phases.

The infobox UI component (jsx) also is a banner component now, because
the text alters often per banner.

Banner_ctrl should call the banner presenter with an instance of the banner to show,
the banner presenter takes care of how the banner will be shown on the page.
The banners include a BannerTransition. The BannerTransition contains how the banner will slide in the page.