Move impression counts to BannerPresenter
BannerPresenter now has a new dependency: the ImpressionCounter.
BannerPresenter increases the impressions when its displays the banner.

This solution is more explicit than the solution on wikipedia.de where
we used two instances of the impression counter (sharing state via the
browser's LocalStorage): One in the entry point to increate the counter
and one in the Form to access the counters.