Add FollowupTransition component
The FollowupTransition is a second transition that takes place when
displaying a full page banner that replaces a mini banner. The mini
banner is hidden and the full page page slides in from the top.

The second important feature of the FollowupTransition component is its
interaction with the surrounding space: The downwards animation of the
banner must first fill the space left by the vanished mini banner, then
it must look like it's "pushing down" the rest of the page, while
maintaining an overall smooth animation style.

This commit introduces the `CssTransition` class that allows to
parameterize transitions. While usually transitions have an
`ease-in-out` easing, the two-phase transition must be split into
`ease-in` and `ease-out` and synchronized between the transition
component and the `skinAdjuster`. For this feature, we'Re exposing the
`skinAdjuster` as a property and moving the initial adjustment to the
`BannerTransition` component.
Added the transition parameter (with defaults) to all old banners. The
skin animation of the old banners might be different from the
jQuery-based animation of the banner itself, but since the old banners
will be deleted, this change only keeps them working.