Fix mobile banners
Safari mobile crashes on close when the Minerva SkinAdjuster moves the
banner element to the top of the DOM.

This change adapts the Transition elements to be sgtill absoluetly
positioned, but inside a relatively positioned parent element,
whose "top" position is changed by skinAdjuster. To keep the 2-phase
animation, the FollowupTransition must have a negative top offset that
corresponds to the positive top offset of its parent element. That way,
it will always stay at position 0. It also needs to incorporate the
previous offset of the mini banner.

Renamed "frbanner" to "fullpage-banner" to properly show and hide the 2
banners depending on state and to have more semantic names.