Improve custom amount handling
These changes have the effect of restoring a previously selected custom
amount selection if the user focuses on the custom amount selection
field and then tabs out of the field without entering a value.

WARNING: This is only implemented as a proof-of-concept for the German
Desktop banner for wp.org

WARNING: When the user clicks on a different amount, the old selected
amount will be selected briefly before the change occurs.