Improve custom amount handling
This change prevents an error message from popping up when the user
focuses the custom amount field and then selects a different amount or
leaves the field via tab. The "Custom Amount" *option* (visible as a
radio button in some banners) stays selected, but the internal state
will still be "unset".

This change also improves the display of the Euro sign in the
SelectCustomAmount component: it only displays the Euro sign when the
amount field has focus or when it has a non-empty value.