Add Travis check for development banner parameters
This commit adds a CI check that throws an error when a pull
request contains a call to createCampaignParameters with arguments.

Background: To override the banner parameters on wikipedia.de with
development ones, you need to call createCampaignParameters with the
argument `new DevCampaignParameters`, while a call without the parameter
will use the parameters from the banner server.