Use English use of funds in English banners
Use English version of use of funds file from the content repository in
the English banners (desktop, iPad, mobile).

Change the image URL for the flow of funds to be inserted from the
content file, so we can change the English one, containing "Ihre
Spende".

Add `font-variant-numeric: tabular-nums;` to company numbers to make the
digits line up better.