Use new slider library "Splider"
The old one, Flickity, threw errors on Firefox, so we're replacing it.
This commit only changes the library for the desktop banner, subsequent
banners should move `Slides.jsx` and the Chevron components from
`desktop/components/ui` into `shared/components/ui` or encapsulate the
functionality into a hook (and use different markup, with and without
chevrons).

This change also moves the banner components into the components
subdirectory, for consistency and better macOS compatibility.

This is for https://phabricator.wikimedia.org/T268783 (error ticket) and
https://phabricator.wikimedia.org/T269104 (campaign ticket).

Co-authored-by: Abban <abban.dunne@wikimedia.de>