Fix FollowupTransition behavior
When the transition duration had a small value, the duration for the 2nd
phase was zero, preventing the onTransitionEnd event to fire (because
there was no transition, only immediate value change). Now we're
checking that condition and change the state to FINISHED immediately if
needed.

We also move the call to `onFinish` to a later stage - now it's called
when the component has finished rendering after the state has changed to
FINISHED, instead of calling it immediately after the call to `setState`.
This enables callbacks to call scroll targets inside the rendered
component.

The third bugfix is to ignore any transitionEnd events that don't
orginate from the component and that would trigger unwanted state
changes.

Adapted the mobile VAR banner to work correctly with the new behaviors.