Adapt coding style to new wikimedia CS
Autofixed JSDoc issues
Removed shadowed variable names (function params having the same name as
a previously defined variable).

Importing classes into files where they are used as parameters types in
JSDoc comments.
I'm not happy with this solution (because the IDE shows the import as unused),
but at the moment it's the only way I know to make the linter not throw
a `jsdoc/no-undefined-types` error.

Deleted the old progress_tree, we've not not used that code in the last
campaign.