Generate form query and action in a hook
Add `useFormAction` hook to dynamically generate the form URL and
centralize the query string. This is important because we're moving the
impression count data to the query string (so the fundraising
application frontend code can pick it up without a server round trip).

Currently, the query string is static. If we need it to be reactive, we
can extend useFormAction hook with a `setQuery` function in the future
that has the same signature as useFormAction.

We did not adapt the desktop banner to avoid merge conflicts. This needs
to be done for the next banner.

Co-Authored-By: Abban Dunne <abban.dunne@wikimedia.de>
Co-Authored-By: Corinna Hillebrand <corinna.hillebrand@wikimedia.de>