Fix impression counts
The banner presenter passes a plain javascript object
that initialises the impression counting values in the
banners. However, it doesn't increment them before passing
them and as the values are not reactive the banners were
one increment behind what they should be.

This is a somewhat temporary fix for that.