Implement VAR C21_WMDE_Test_21
Add sentence for remaining donors. To avoid duplication of components
(`BannerText_var` and `Slides_var`), the sentence has an empty
translation in CTRL. This should be removed in the next test that takes
CTRL as a basis.

Add calculation and test to CampaignProjection class.

Add new property `averageAmountPerDonation` to campaignParameters class/interface.
ATTENTION: The new property is a potential BC break (when it's missing
on the server). If it's not needed/used then it's fine.
ATTENTION: `averageAmountPerDonation` is not (yet) used in the BannerText/Slides components.
It's still hardcoded there because it needs proper localized number formatting with currency symbol,
which is out of scope for this ticket.

Clean up translations, both shared and for desktop.