Adapt mediawiki_text_wrapper for webpack 5
It's no longer allowed to manuipulate the compilation asset object,
instead we need to use the emitAsset and deleteAsset methods.

Also, the emit hook is not the right place to tap into processAssets,
using the compilation hook works.

This gets rid of the deprecation warning when building.

This is for https://phabricator.wikimedia.org/T311881