Fix IE11 weirdness with svg icons
So this is a weird one and I can't understand
why it's happening. IE11 doesn't like the error
icons that appear in the form validation. It likes
the exact same ones in other banners but for some
reason not on WPDE desktop. In order to make it
happy uploaded the SVG to commons and referenced
it as a background.