Add Git Branch to info
Restructure webpack.config to load campaigns and name of current git
branch asynchronously.

Output git branch in Dashboard. The SSH command assumes and alias and
assumes that the banner screenshots tool is in the main directory (which
it isn't at the moment). This is for security purposes, we don't want to
expose host or path names.

Fix code style